add: hold right click to use inverted brush
